refactor(settings): move to react
