+++ app/edge/components/useIntervalOptions.ts
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { useState, useEffect } from 'react';
 
-import { useSettings } from '@/portainer/settings/queries';
+import { useSettings } from '@/react/portainer/settings/queries';
 
 type Option = {
   label: string;
