+++ app/react/portainer/registries/ListView/DefaultRegistry/DefaultRegistryDomain.tsx
@@ -1,5 +1,6 @@
 import clsx from 'clsx';
-import { usePublicSettings } from 'Portainer/settings/queries';
+
+import { usePublicSettings } from '@/react/portainer/settings/queries';
 
 export function DefaultRegistryDomain() {
   const settingsQuery = usePublicSettings({
