+++ app/react/portainer/settings/EdgeComputeView/AutomaticEdgeEnvCreation/AutomaticEdgeEnvCreation.tsx
@@ -2,11 +2,12 @@ import { useMutation } from 'react-query';
 import { useEffect } from 'react';
 
 import { generateKey } from '@/portainer/environments/environment.service/edge';
-import { useSettings } from '@/portainer/settings/queries';
 import { EdgeScriptForm } from '@/react/edge/components/EdgeScriptForm';
 import { commandsTabs } from '@/react/edge/components/EdgeScriptForm/scripts';
 
 import { Widget, WidgetBody, WidgetTitle } from '@@/Widget';
+
+import { useSettings } from '../../queries';
 
 import { AutoEnvCreationSettingsForm } from './AutoEnvCreationSettingsForm';
