+++ app/react/sidebar/SettingsSidebar.tsx
@@ -7,7 +7,7 @@ import {
   FileText,
 } from 'react-feather';
 
-import { usePublicSettings } from '@/portainer/settings/queries';
+import { usePublicSettings } from '@/react/portainer/settings/queries';
 import {
   FeatureFlag,
   useFeatureFlag,
