+++ app/setup-tests/server-handlers.ts
@@ -9,7 +9,7 @@ import { EnvironmentGroup } from '@/portainer/environment-groups/types';
 import { Tag } from '@/portainer/tags/types';
 import { StatusResponse } from '@/portainer/services/api/status.service';
 import { createMockTeams } from '@/react-tools/test-mocks';
-import { PublicSettingsResponse } from '@/portainer/settings/types';
+import { PublicSettingsResponse } from '@/react/portainer/settings/types';
 import { UserId } from '@/portainer/users/types';
 
 import { azureHandlers } from './setup-handlers/azure';
